+++ src/main/java/osadsakana/utilitiesforprogrammers/mixin/MinecraftMixin.java
@@ -42,7 +42,7 @@ public class MinecraftMixin {
         }
     }
 
-    @Inject(method = "pickBlock", at = @At("HEAD"), cancellable = true)
+    @Inject(method = "pickBlockOrEntity", at = @At("HEAD"), cancellable = true)
     private void ufp$blockPickBlock(CallbackInfo ci) {
         if (ToggleState.isFrozen()) {
             ci.cancel();